♻️ refactor [#240] : 상품 리펙터링, 재고 엔티티 생성
 - 상품에서 재고 관련 기능을 재고 엔티티로 분리
