+++ product-service/src/main/java/com/sparta/product/domain/model/Product.java
@@ -36,17 +36,11 @@ public class Product extends BaseEntity {
     @Column(name = "product_price", nullable = false)
     private BigDecimal price;
 
-    @Column(name = "product_quantity", nullable = false)
-    private Integer quantity;
-
     @Column(name = "is_display", nullable = false)
     private boolean isDisplay;
 
     @Column(name = "company_id", nullable = false)
     private UUID companyId;
-
-    @Column(name = "hub_id", nullable = false)
-    private UUID hubId;
 
 
     /**
@@ -57,10 +51,8 @@ public class Product extends BaseEntity {
                 .name(requestDto.name())
                 .description(requestDto.description())
                 .price(requestDto.price())
-                .quantity(requestDto.quantity())
                 .isDisplay(requestDto.isDisplay())
                 .companyId(requestDto.companyId())
-                .hubId(requestDto.hubId())
                 .build();
     }
 
@@ -77,48 +69,15 @@ public class Product extends BaseEntity {
     }
 
 
-    /**
-     *  상품 수정(재고 감소)
-     */
-    public void decreaseQuantity(Integer quantity) {
-        validateDecreaseQuantity(quantity);
-        this.quantity = this.quantity - quantity;
-    }
-
-
-
-    /**
-     *  상품 수정(재고 증가)
-     */
-    public void increaseQuantity(Integer quantity) {
-        this.quantity = this.quantity + quantity;
-    }
-
-
-    private void validateDecreaseQuantity(Integer quantity) {
-        if (quantity == null || quantity < 30) {
-            throw new IllegalArgumentException("최소 30개 이상 요청해야 합니다.");
-        }
-        if (this.quantity < quantity) {
-            throw new IllegalStateException("재고가 부족합니다.");
-        }
-    }
-
-
-
 
     // DTO -> Entity 변환 메서드
     public static Product of(CreateProductResponseDto responseDto) {
         return Product.builder()
                 .name(responseDto.name())
                 .description(responseDto.description())
-                .quantity(responseDto.quantity())
                 .price(responseDto.price())
                 .isDisplay(responseDto.isDisplay())
                 .companyId(responseDto.companyId())
-                .hubId(responseDto.hubId())
                 .build();
     }
-
-
 }
